planet and moon gen is done - atmospheres

+++ mysite/solapi/models.py
@@ -51,14 +51,14 @@ class Star(models.Model):
 
 
 class Planet(models.Model):
-    name = models.CharField(max_length=100)
-    seed = models.CharField(max_length=100)
-    planetType = models.CharField(max_length=100)
-    mass = models.IntegerField()
-    radius = models.IntegerField()
-    gravity = models.IntegerField()
-    tilt = models.IntegerField()
-    axis = models.IntegerField()
+    name = models.CharField(max_length=100) # the name of the planet
+    seed = models.CharField(max_length=100) # the random seed
+    planetType = models.CharField(max_length=100) # ice or gas or rock
+    mass = models.FloatField() # the mass of the planet Units: kg
+    radius = models.FloatField() # radius of the planet Units: AU
+    gravity = models.FloatField() # gravity or little g Units: kgms^2 i think
+    tilt = models.FloatField() # the tilt the planet of the spin from the sun Units: deg
+    axis = models.FloatField() # the tilt of the orbetal plane Units: deg 
 
     atmosphere = models.OneToOneField(Atmosphere, on_delete=models.CASCADE)
     orbit = models.OneToOneField(Orbit, on_delete=models.CASCADE)
